return PID weighted and normalised for when both sources are present
edited Conor's script so that it returns the PID weighted for the top
two lines, which are the ones of interest (both
"sources" (i.e. selection coefficients) present for target =
0 (extinction) and target = 1 (allele persists). Also normalises the
total information (U1 + U2 + R + C) so it sums to 1, so that we can
talk about apportioning the function metric (rather than the amount of
information in bits/nats, etc.).