remove number_generations and gen_env_2
now calculate as t -> inf

just a note that I might need to be careful which parameter values I
use for DSE as it can display negative frequency dependency and it will
be possible to have some parameter value combinations that take
forever to reach one of the absorbing states (e.g. strong
overdominance with extremely high population sizes). In fact, to be
safe, I think I'll hardcode a max number of generations for the DSE
model after which it exits the while loop and records that the allele
still persists